Refactor parse_likes : move dichotomic search

+++ include/parser.h
@@ -132,4 +132,4 @@ UserData *to_user_oriented(MovieData *data);
  * @param movie_data Data of movies to find corresponding identifiers.
  * @return `uint16_t*` An array containing the identifiers of liked movies.
  */
-uint16_t* parse_likes(char *filename, MovieData *movie_data);
+uint* parse_likes(char *filename, MovieData *movie_data);
